Fix build on clang.
(It's complaining about possible truncations and seems to be
extraordinarily naive about it)

+++ src/include/bitstring_inline.h
@@ -383,14 +383,14 @@ __bit_clear_end(uint8_t *bitf, uint64_t numentries, uint8_t width)
 
 	firstbit = numentries * width;
         byte = __bit_byte(firstbit);
-        mask = __bit_mask(firstbit);
+        mask = (uint8_t)__bit_mask(firstbit);
 
         /* If mask is the first bit of the next byte, we don't need to do anything. */
         if (mask == 0x01)
             return;
 
         /* Convert e.g. 0b01000000 to 0b01111111 and use the resulting mask to clear. */
-        mask = (mask << 1) - 1;
+        mask = (uint8_t)((uint16_t)mask << 1) - 1;
         bitf[byte] &= ~mask;
 }
 
@@ -406,13 +406,13 @@ __bit_end_is_clear(const uint8_t *bitf, uint64_t numentries, uint8_t width)
 
 	firstbit = numentries * width;
         byte = __bit_byte(firstbit);
-        mask = __bit_mask(firstbit);
+        mask = (uint8_t)__bit_mask(firstbit);
 
         /* If mask is the first bit of the next byte, there's nothing to check. */
         if (mask == 0x01)
             return (true);
 
         /* Convert e.g. 0b01000000 to 0b01111111 and read with the resulting mask. */
-        mask = (mask << 1) - 1;
+        mask = (uint8_t)(((uint16_t)mask << 1) - 1);
         return ((bitf[byte] & mask) == 0);
 }
